JS frontend will cast any Numeric form data to it's proper type
This is a breaking change.
All form fields data will be casted to Numeric types before serializing and
sending to backend. That is
 - checkboxes will be casted to boolean true/false
 - numeric data will be casted to int/float
 - all NaN data will be sent as-is like text strings

On the backend side it is possible to retrieve key:val pairs as JsonVarians casted to any type required,
used in bool comparison, assigned to int variables, etc...

+ option in drop-down lists is a template now

Signed-off-by: Emil Muratov <gpm@hotplug.ru>