add new method Interface::make_new_object()
It creates a new JsonObject at the end of current block and returns a referencing object.
Since ArduinoJson 7 it is not required to fit into a limit of fixed size JsonDocument,
so creating an elements in-place appending to Interface's document could help to avoid useless data copy.

Another benefit - it's more simple to create any arbitrary objects without modifying lib's code and
adding specific methods to Interface class.